refactor(utils): Move utils/get-asset* to assets/
