+++ packages/utils/src/assets/__tests__/get-asset-name.js
