+++ packages/utils/src/assets/get-asset-name.js
