+++ packages/utils/src/assets/get-assets-metrics.js
