+++ packages/utils/src/assets/index.js
@@ -1,3 +1,5 @@
 export * from './calculate-totals';
 export * from './calculate-cache-invalidation';
 export * from './file-types';
+export * from './get-asset-name';
+export * from './get-assets-metrics';
